refactor: 랭킹 산정 방식 Dense로 변경
