+++ backend/bottari/src/main/java/com/bottari/apiworkshop/Cursor.java
@@ -1,0 +1,11 @@
+package com.bottari.apiworkshop;
+
+// TODO: nextCursor로 변경 후, 이를 인코딩
+record Cursor(
+        int limit,
+        String lastId,
+        boolean hasNext,
+        int lastRank,
+        Long lastCount
+) {
+}
